refactor poll deferred indexes
