+++ internal/api/gsi.go
@@ -1,5 +1,16 @@
 package api
 
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"net/url"
+	"time"
+
+	internalerrors "github.com/couchbasecloud/terraform-provider-couchbase-capella/internal/errors"
+)
+
 // IndexDDLRequest is a request to run an index DDL statement.
 type IndexDDLRequest struct {
 	// Definition The index DDL statement.  This can be a CREATE/DROP/ALTER/BUILD statement.
@@ -45,4 +56,127 @@ type ListIndexDefinitionsResponse struct {
 // IndexBuildStatusResponse is the build status for an index.
 type IndexBuildStatusResponse struct {
 	Status string `json:"status"`
+}
+
+// TODO: get build status for all indexes in 1 API call
+// TODO: combine WatchIndexes and EnsureIndexesAreCreated
+
+type Options struct {
+	Host       string
+	OrgId      string
+	ProjectId  string
+	ClusterId  string
+	Bucket     string
+	Scope      string
+	Collection string
+}
+
+func WatchIndexes(
+	ctx context.Context, indexes []string, exec func(cfg EndpointCfg) (response *Response, err error),
+	options Options,
+) error {
+
+	const (
+		maxDuration = 20
+		timeout     = time.Minute * 60 // 60 min is arbitrary
+	)
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
+	attempt := 0
+	timer := time.NewTimer((1 << attempt) * time.Minute)
+
+	for {
+		select {
+		case <-ctx.Done():
+			return internalerrors.ErrLongIndexBuildTime
+		case <-timer.C:
+			for i := 0; i < len(indexes); {
+				url := fmt.Sprintf(
+					"%s/v4/organizations/%s/projects/%s/clusters/%s/queryService/indexBuildStatus/%s?bucket=%s&scope=%s&collection=%s",
+					options.Host,
+					options.OrgId,
+					options.ProjectId,
+					options.ClusterId,
+					url.QueryEscape(indexes[i]),
+					options.Bucket,
+					options.Scope,
+					options.Collection,
+				)
+
+				cfg := EndpointCfg{Url: url, Method: http.MethodGet, SuccessStatus: http.StatusOK}
+				response, err := exec(cfg)
+				if err != nil {
+					break
+				}
+
+				status := IndexBuildStatusResponse{}
+				if err = json.Unmarshal(response.Body, &status); err != nil {
+					return err
+				}
+
+				if status.Status != "Ready" {
+					break
+				}
+			}
+
+			// exponential backoff upto a max of 20 min.
+			d := min(maxDuration, 1<<attempt)
+			timer.Reset(time.Duration(d) * time.Minute)
+		}
+	}
+}
+
+func EnsureIndexesAreCreated(
+	indexes []string, exec func(cfg EndpointCfg) (response *Response, err error),
+	options Options,
+) error {
+	const maxAttempts = 5
+	attempts := 0
+
+	for i := 0; i < len(indexes); {
+		url := fmt.Sprintf(
+			"%s/v4/organizations/%s/projects/%s/clusters/%s/queryService/indexBuildStatus/%s?bucket=%s&scope=%s&collection=%s",
+			options.Host,
+			options.OrgId,
+			options.ProjectId,
+			options.ClusterId,
+			url.QueryEscape(indexes[i]),
+			options.Bucket,
+			options.Scope,
+			options.Collection,
+		)
+
+		cfg := EndpointCfg{Url: url, Method: http.MethodGet, SuccessStatus: http.StatusOK}
+		response, err := exec(cfg)
+		// retry even on 404 because the index may not have been created yet
+		// so wait and check again
+		if err != nil {
+			attempts++
+			if attempts > maxAttempts {
+				return internalerrors.ErrMaxAttemptsExceeded
+			}
+			time.Sleep(10 * time.Second)
+			continue
+		}
+
+		status := IndexBuildStatusResponse{}
+		if err = json.Unmarshal(response.Body, &status); err != nil {
+			return err
+		}
+
+		if status.Status != "Created" {
+			attempts++
+			if attempts > maxAttempts {
+				return internalerrors.ErrMaxAttemptsExceeded
+			}
+			time.Sleep(10 * time.Second)
+			continue
+		}
+
+		i++
+	}
+
+	return nil
 }
